separate AbstractEnv and ServerEnv into two files
